Deal with prohibited explicit type conversions
Solidity 0.8 prohibited some kinds of explicit
conversions which were possible in former versions.

For `CheckBitcoinSigs.sol`, we make the conversion from `uint256` to `address` working by adding an intermediate
`uint160` step.

For `TypedMemView.sol` the situation is much more
complicated as it contains mutliple conversions
which can't be easily replaced. Since this contract
is not needed by our client code, we remove it
for now. In case it's needed we can restore it
in future and invest some work needed to port
the code to Solidity 0.8.