Repair `BTCUtils.extractHash` tests
The `extractHash` function does not check
for overflow and rely on `if` guards to
detect abnormal situations and return
an empty hex. The corresponding test
of non-happy path expected that empty
hex. However, Solidity 0.8 adds checked
math so `extractHash` will now revert
instead of returning empty hex upon
overflow. This forces us to change the
test scenario to expect a revert instead
of empty hex.