Add workflows for running tests and publishing NPM package
Adding GH Actions workflow that compiles contracts and runs unit tests.
Workflow will be executed every time a PR is created or there's a change
merged to `main` branch. It will also execute every night at midnight
and every time someone dispatches it manually.
Also adding a workflow which will publish a package to NPM registry.
Workflow will execute every time a change in the `solidity/package.json`
file is pushed to a branch which name starts with
`releases/mainnet/solidity/` (or when someone dispatches workflow
manually). Published package will be versioned exactly as it is secfied
in the `package.json` file. The publish will fail if the package name
and version combination already exists in the NPM registry.